refactor(kernel): move evidence discovery out of request_handlers
Collect server-log scanning, trace resolution, and specialist CSV lookup in
kernel_evidence.py. Precision resolution lives in kernel_context without the
runtime-workload exception fallback. Drop duplicate resolver tests and dead
wrapper coverage that the lane projections already own.
