+++ src/hyperloom/orchestrator/kernel/tests/test_forge_server_log_evidence.py
@@ -15,6 +15,7 @@ from pathlib import Path
 
 import pytest
 
+from hyperloom.orchestrator.kernel import kernel_evidence as ke
 from hyperloom.orchestrator.kernel import request_handlers as rh
 
 # A real aiter hit line, copied from a fleet server.log.
@@ -45,29 +46,29 @@ def _log(path: Path, text: str) -> Path:
 
 class TestEvidenceDetection:
     def test_a_hit_line_is_evidence(self, tmp_path):
-        assert rh._log_has_aiter_evidence(_log(tmp_path / "a.log", HIT.format(m=128)))
+        assert ke._log_has_aiter_evidence(_log(tmp_path / "a.log", HIT.format(m=128)))
 
     def test_a_miss_line_is_also_evidence(self, tmp_path):
         # A miss still proves the process routed a GEMM through aiter, which is
         # what makes the log a usable shape source.
-        assert rh._log_has_aiter_evidence(_log(tmp_path / "a.log", MISS.format(m=128)))
+        assert ke._log_has_aiter_evidence(_log(tmp_path / "a.log", MISS.format(m=128)))
 
     def test_a_quiet_log_is_not_evidence(self, tmp_path):
-        assert not rh._log_has_aiter_evidence(_log(tmp_path / "a.log", QUIET))
+        assert not ke._log_has_aiter_evidence(_log(tmp_path / "a.log", QUIET))
 
     def test_a_missing_file_is_not_evidence(self, tmp_path):
-        assert not rh._log_has_aiter_evidence(tmp_path / "nope.log")
+        assert not ke._log_has_aiter_evidence(tmp_path / "nope.log")
 
     def test_a_marker_straddling_a_chunk_boundary_is_still_found(self, tmp_path, monkeypatch):
         # Fleet logs are ~17MB, so the scan is chunked; the overlap must cover a
         # marker split across two reads.
-        monkeypatch.setattr(rh, "_LOG_SCAN_CHUNK", 16)
+        monkeypatch.setattr(ke, "_LOG_SCAN_CHUNK", 16)
         text = "x" * 10 + HIT.format(m=128)
-        assert rh._log_has_aiter_evidence(_log(tmp_path / "a.log", text))
+        assert ke._log_has_aiter_evidence(_log(tmp_path / "a.log", text))
 
     def test_evidence_late_in_a_large_log_is_found(self, tmp_path):
         text = ("noise line\n" * 200_000) + HIT.format(m=99)
-        assert rh._log_has_aiter_evidence(_log(tmp_path / "a.log", text))
+        assert ke._log_has_aiter_evidence(_log(tmp_path / "a.log", text))
 
     def test_a_moe_only_log_is_evidence_too(self, tmp_path):
         # The resolved log is not only a dense-shape source: kernelforge's router
@@ -75,17 +76,17 @@ class TestEvidenceDetection:
         # [fused_moe] lines. Rejecting a log that has those but no dense
         # "shape is M:" line would blind the router on exactly the MoE models
         # this lane runs against.
-        assert rh._log_has_aiter_evidence(_log(tmp_path / "a.log", MOE))
-        assert rh._log_has_aiter_evidence(_log(tmp_path / "b.log", "Mxfp4 MoE backend selected\n"))
+        assert ke._log_has_aiter_evidence(_log(tmp_path / "a.log", MOE))
+        assert ke._log_has_aiter_evidence(_log(tmp_path / "b.log", "Mxfp4 MoE backend selected\n"))
 
     def test_a_moe_only_log_yields_no_dense_tokens(self, tmp_path):
         # ...and it must not invent any: --tokens comes from dense M only.
-        assert rh._tokens_from_serving_log(_log(tmp_path / "a.log", MOE)) == ""
+        assert ke._tokens_from_serving_log(_log(tmp_path / "a.log", MOE)) == ""
 
     def test_a_dispatch_line_at_m_zero_is_still_evidence(self, tmp_path):
         # The M counter skips 0, so an evidence check derived from its output
         # read this log as silent.
-        assert rh._log_has_aiter_evidence(_log(tmp_path / "a.log", HIT.format(m=0)))
+        assert ke._log_has_aiter_evidence(_log(tmp_path / "a.log", HIT.format(m=0)))
 
 
 class TestSelection:
@@ -94,7 +95,7 @@ class TestSelection:
         _log(ws / "server.log", QUIET)
         good = _log(tmp_path / "runs" / "baseline" / "h0" / "warmup_round" / "b0" / "server.log", HIT.format(m=256))
 
-        picked = rh._resolve_forge_server_log(_State(current_best={"workspace": str(ws)}), tmp_path)
+        picked = ke._resolve_forge_server_log(_State(current_best={"workspace": str(ws)}), tmp_path)
 
         assert picked == str(good)
 
@@ -103,7 +104,7 @@ class TestSelection:
         mine = _log(ws / "server.log", HIT.format(m=256))
         _log(tmp_path / "runs" / "baseline" / "h0" / "warmup_round" / "b0" / "server.log", HIT.format(m=1))
 
-        assert rh._resolve_forge_server_log(_State(current_best={"workspace": str(ws)}), tmp_path) == str(mine)
+        assert ke._resolve_forge_server_log(_State(current_best={"workspace": str(ws)}), tmp_path) == str(mine)
 
     def test_the_warmup_sibling_search_skips_quiet_logs(self, tmp_path):
         # server.log lives in warmup_round while current_best points at
@@ -119,7 +120,7 @@ class TestSelection:
         os.utime(older, (1, 1))  # make the quiet one strictly newer
 
         assert newest.stat().st_mtime > older.stat().st_mtime
-        assert rh._resolve_forge_server_log(_State(current_best={"workspace": str(ws)}), tmp_path) == str(older)
+        assert ke._resolve_forge_server_log(_State(current_best={"workspace": str(ws)}), tmp_path) == str(older)
 
     def test_the_runs_fallback_skips_quiet_logs(self, tmp_path):
         import os
@@ -129,13 +130,13 @@ class TestSelection:
         os.utime(good, (1, 1))
 
         assert newer_quiet.stat().st_mtime > good.stat().st_mtime
-        assert rh._resolve_forge_server_log(_State(), tmp_path) == str(good)
+        assert ke._resolve_forge_server_log(_State(), tmp_path) == str(good)
 
     def test_no_candidate_with_evidence_returns_empty_and_says_why(self, tmp_path, caplog):
         _log(tmp_path / "runs" / "baseline" / "h0" / "warmup_round" / "b0" / "server.log", QUIET)
 
         with caplog.at_level("WARNING"):
-            assert rh._resolve_forge_server_log(_State(), tmp_path) == ""
+            assert ke._resolve_forge_server_log(_State(), tmp_path) == ""
 
         # "no log at all" and "logs exist but are silent" are different
         # problems; only the second is actionable.
@@ -144,7 +145,7 @@ class TestSelection:
     def test_no_logs_at_all_is_quiet(self, tmp_path, caplog):
         (tmp_path / "runs").mkdir()
         with caplog.at_level("WARNING"):
-            assert rh._resolve_forge_server_log(_State(), tmp_path) == ""
+            assert ke._resolve_forge_server_log(_State(), tmp_path) == ""
         assert "AITER_LOG_TUNED_CONFIG" not in caplog.text
 
     def test_baseline_is_consulted_when_current_best_is_quiet(self, tmp_path):
@@ -156,24 +157,24 @@ class TestSelection:
 
         state = _State(current_best={"workspace": str(quiet)}, last_baseline={"workspace": str(base)})
 
-        assert rh._resolve_forge_server_log(state, tmp_path) == str(real)
+        assert ke._resolve_forge_server_log(state, tmp_path) == str(real)
 
 
 class TestTokensFromServingLog:
     def test_the_observed_m_values_come_back_sorted(self, tmp_path):
         text = "".join(HIT.format(m=m) for m in (512, 128, 15842))
-        assert rh._tokens_from_serving_log(_log(tmp_path / "a.log", text)) == "128,512,15842"
+        assert ke._tokens_from_serving_log(_log(tmp_path / "a.log", text)) == "128,512,15842"
 
     def test_a_miss_line_counts_too(self, tmp_path):
         text = MISS.format(m=64) + HIT.format(m=32)
-        assert rh._tokens_from_serving_log(_log(tmp_path / "a.log", text)) == "32,64"
+        assert ke._tokens_from_serving_log(_log(tmp_path / "a.log", text)) == "32,64"
 
     def test_the_most_frequent_m_values_win_the_budget(self, tmp_path):
         # 1 appears once, 2..4 appear three times each; with a budget of 3 the
         # rare one is dropped. Tuning where the model spends its time beats
         # tuning the largest M it ever reached.
         text = HIT.format(m=1) + "".join(HIT.format(m=m) * 3 for m in (2, 3, 4))
-        assert rh._tokens_from_serving_log(_log(tmp_path / "a.log", text), limit=3) == "2,3,4"
+        assert ke._tokens_from_serving_log(_log(tmp_path / "a.log", text), limit=3) == "2,3,4"
 
     def test_uniform_counts_do_not_starve_the_prefill_end(self, tmp_path):
         # Regression: a serving warmup sweeps every M about equally often, so
@@ -184,7 +185,7 @@ class TestTokensFromServingLog:
         # M x40), and both missed on 16384/24576/32768 and 57344/65536.
         ms = [1, 2, 4, 8, 16, 32, 64, 128, 256, 512, 1024, 2048, 8192, 16384, 24576, 32768]
         text = "".join(HIT.format(m=m) * 4 for m in ms)
-        got = rh._tokens_from_serving_log(_log(tmp_path / "a.log", text), limit=8)
+        got = ke._tokens_from_serving_log(_log(tmp_path / "a.log", text), limit=8)
         picked = {int(t) for t in got.split(",")}
         assert {24576, 32768} <= picked, got
         assert len(picked) == 8, got
@@ -192,23 +193,23 @@ class TestTokensFromServingLog:
         assert len([m for m in picked if m <= 256]) >= 4, got
 
     def test_a_quiet_log_yields_nothing(self, tmp_path):
-        assert rh._tokens_from_serving_log(_log(tmp_path / "a.log", QUIET)) == ""
+        assert ke._tokens_from_serving_log(_log(tmp_path / "a.log", QUIET)) == ""
 
     def test_a_missing_log_yields_nothing(self, tmp_path):
-        assert rh._tokens_from_serving_log(tmp_path / "nope.log") == ""
+        assert ke._tokens_from_serving_log(tmp_path / "nope.log") == ""
 
     def test_the_result_is_what_forge_accepts(self, tmp_path):
         # forge parses --tokens as int(t) for t in value.split(","); round-trip
         # through the normaliser must not change it.
         text = "".join(HIT.format(m=m) for m in (7, 4096))
-        raw = rh._tokens_from_serving_log(_log(tmp_path / "a.log", text))
+        raw = ke._tokens_from_serving_log(_log(tmp_path / "a.log", text))
         assert rh._normalize_tokens(raw) == raw
         assert [int(t) for t in raw.split(",")] == [7, 4096]
 
     def test_m_values_are_read_across_chunk_boundaries(self, tmp_path, monkeypatch):
-        monkeypatch.setattr(rh, "_LOG_SCAN_CHUNK", 24)
+        monkeypatch.setattr(ke, "_LOG_SCAN_CHUNK", 24)
         text = "".join(HIT.format(m=m) for m in (11, 22, 33))
-        assert rh._tokens_from_serving_log(_log(tmp_path / "a.log", text)) == "11,22,33"
+        assert ke._tokens_from_serving_log(_log(tmp_path / "a.log", text)) == "11,22,33"
 
 
 class TestShapeManifestResolution:
@@ -222,42 +223,13 @@ class TestShapeManifestResolution:
             path.write_text("{}", encoding="utf-8")
         os.utime(old, (1, 1))
 
-        assert rh._resolve_trace_shape_manifest(_State(), tmp_path) == str(new)
+        assert ke._resolve_trace_shape_manifest(_State(), tmp_path) == str(new)
 
     def test_no_manifest_is_an_empty_string_not_an_error(self, tmp_path):
-        assert rh._resolve_trace_shape_manifest(_State(), tmp_path) == ""
+        assert ke._resolve_trace_shape_manifest(_State(), tmp_path) == ""
 
     def test_a_missing_session_dir_is_survivable(self, tmp_path):
-        assert rh._resolve_trace_shape_manifest(_State(), tmp_path / "gone") == ""
-
-
-class TestTheWrapperForwardsTheNewInputs:
-    @pytest.mark.parametrize(
-        "key,flag",
-        [
-            ("shapes_manifest", "--shapes-manifest"),
-        ],
-    )
-    def test_a_populated_field_reaches_the_forge_cli(self, key, flag):
-        import importlib.util
-
-        spec = importlib.util.spec_from_file_location(
-            "_fgt",
-            Path(rh.__file__).parents[2] / "agents" / "kernel" / "tools" / "forge_gemm_tuning.py",
-        )
-        mod = importlib.util.module_from_spec(spec)
-        spec.loader.exec_module(mod)
-
-        base = {
-            "model_path": "/m",
-            "framework": "sglang",
-            "precision": "bf16",
-            "output_dir": "/o",
-        }
-        assert flag not in mod._build_cmd(base)
-
-        cmd = mod._build_cmd({**base, key: "/tmp/x.json"})
-        assert cmd[cmd.index(flag) + 1] == "/tmp/x.json"
+        assert ke._resolve_trace_shape_manifest(_State(), tmp_path / "gone") == ""
 
 
 class _StopHere(Exception):
@@ -309,7 +281,7 @@ class TestTokensAreDerivedBeforeTheMoeCsvIsBuilt:
         with pytest.raises(_StopHere):
             await rh._run_forge_gemm_tuning(payload, session_dir=tmp_path)
 
-        assert seen["tokens"] == rh._tokens_from_serving_log(log_path)
+        assert seen["tokens"] == ke._tokens_from_serving_log(log_path)
         # Not the ``[1]`` fallback, and not a single value of any kind.
         assert len(seen["tokens"].split(",")) > 1
